[IMP] Run tests on all supported Odoo versions
Related to #1.

This commit adds Odoo versions 14 and 15 to be run during the test
pipeline. It also imports annotations from __future__ in order to use
type hinting with older Python versions found in Odoo setups.

Instructions on running the test suite with different odoo versions have
also been added to the README.